calcuate change in theta(the gradient)
um of the product of X and the "errors vector", scaled by alpha and 1/m.
Since X is (m x n), and the error vector is (m x 1), and the result you want is the same size as theta (which is (n x 1), you need to transpose X before you can multiply it by the error vector.

The vector multiplication automatically includes calculating the sum of the products. So you do not need to use the "sum()" function separately.

When you're scaling by alpha and 1/m, be sure you use enough sets of parenthesis to get the factors correct.